v2.5 release

+++ src/models/dit_3b/normalization.py
@@ -97,11 +97,12 @@ class CustomRMSNorm(nn.Module):
         normalized = input / rms
         
         if self.elementwise_affine:
-            # Convert FP8 weight to BFloat16 for arithmetic operations
+            # Convert FP8 weight to match input dtype for arithmetic operations
             if hasattr(torch, 'float8_e4m3fn'):
                 fp8_types = (torch.float8_e4m3fn, torch.float8_e5m2)
                 if self.weight.dtype in fp8_types:
-                    weight = self.weight.to(torch.bfloat16)
+                    # Use input dtype as target (respects pipeline precision)
+                    weight = self.weight.to(input.dtype)
                     return normalized * weight
                     
             return normalized * self.weight
